feat(mcp-server): unify developer and secops servers behind OAuth scopes
Collapse the developer and SecOps MCP servers into a single `gg-mcp-server`
package. Permissions are now governed entirely by the access token's OAuth
scopes — the `ScopeFilteringMiddleware` already hides tools whose required
scopes the token cannot satisfy, so a single registration list serves both
audiences. Removes the `SERVER_PROFILE` env var and the `ServerProfile`
enum; the host-based `MINIMAL_SCOPES` cap for non-local self-hosted
instances is kept in `Settings.effective_scopes`.

The old `developer_mcp_server` and `secops_mcp_server` packages become
deprecation shims: every module re-exports from `gg_mcp_server` and emits
a `DeprecationWarning`. The `developer-mcp-server` and `secops-mcp-server`
console scripts still work for one release, and the BC wheel builds in the
Dockerfile preserve `secops_mcp_server.http_app:app` for the existing
Helm chart gunicorn target.

Refs: SI-3352
