+++ packages/developer_mcp_server/src/developer_mcp_server/http_app.py
@@ -1,20 +1,13 @@
-"""ASGI application for MCP server over StreamableHTTP.
+"""Deprecated compatibility shim — use ``gg_mcp_server.http_app`` instead."""
 
-This module exports the ASGI application for use with ASGI servers like gunicorn + uvicorn.
-It imports the configured MCP server and exposes its StreamableHTTP application.
-"""
+import warnings
 
-from fastmcp.server.http import create_streamable_http_app
+from gg_mcp_server.http_app import http_app
 
-from developer_mcp_server.server import mcp
-
-# Note: We use StreamableHTTP with json_response=True and stateless_http=True to enable
-# fully stateless operation. This allows horizontal scaling without sticky sessions
-# since no session state is maintained between requests.
-http_app = create_streamable_http_app(
-    server=mcp,
-    streamable_http_path="/mcp",
-    auth=mcp.auth,
-    json_response=True,
-    stateless_http=True,
+warnings.warn(
+    "developer_mcp_server.http_app is deprecated; use gg_mcp_server.http_app instead.",
+    DeprecationWarning,
+    stacklevel=2,
 )
+
+__all__ = ["http_app"]
