+++ packages/developer_mcp_server/src/developer_mcp_server/sse_app.py
@@ -1,16 +1,13 @@
-"""ASGI application for MCP server over HTTP/SSE.
+"""Deprecated compatibility shim — use ``gg_mcp_server.sse_app`` instead."""
 
-Note: This SSE transport requires sticky sessions for horizontal scaling since
-session state is maintained in-memory per worker. For stateless operation,
-use http_app.py instead which uses StreamableHTTP with JSON responses.
-"""
+import warnings
 
-from fastmcp.server.http import create_sse_app
+from gg_mcp_server.sse_app import sse_app
 
-from developer_mcp_server.server import mcp
-
-sse_app = create_sse_app(
-    server=mcp,
-    message_path="/messages/",
-    sse_path="/sse",
+warnings.warn(
+    "developer_mcp_server.sse_app is deprecated; use gg_mcp_server.sse_app instead.",
+    DeprecationWarning,
+    stacklevel=2,
 )
+
+__all__ = ["sse_app"]
