+++ packages/gg_mcp_server/README.md
@@ -1,0 +1,7 @@
+# gg-mcp-server
+
+Unified GitGuardian MCP server. The tools exposed at runtime depend on the
+OAuth scopes granted to the access token used to authenticate — there is no
+longer a separate "developer" or "secops" build.
+
+See the project README at the repository root for installation instructions.
