+++ packages/gg_mcp_server/src/gg_mcp_server/__init__.py
@@ -1,0 +1,3 @@
+"""GitGuardian MCP server (unified)."""
+
+__version__ = "0.1.0"
