+++ packages/secops_mcp_server/src/secops_mcp_server/run.py
@@ -1,66 +1,25 @@
-"""Runtime entry points for the SecOps MCP server.
+"""Deprecated compatibility shim — use ``gg-mcp-server`` instead."""
 
-This module provides different ways to run the MCP server:
-- stdio: Standard input/output transport (default for CLI tools)
-- http: StreamableHTTP transport using uvicorn (for local development)
-"""
+import warnings
 
-import logging
+from gg_mcp_server.run import (
+    run_http_with_uvicorn,
+    run_stdio,
+)
+from gg_mcp_server.run import (
+    run_mcp_server as _run_mcp_server,
+)
 
-from gg_api_core.sentry_integration import init_sentry
-from gg_api_core.settings import get_settings
-
-from secops_mcp_server.server import mcp
-
-logger = logging.getLogger(__name__)
-
-
-def run_stdio():
-    """Run the MCP server over stdio transport.
-
-    This is the default mode for MCP servers, used when the server
-    is invoked as a subprocess by MCP clients like Claude Desktop.
-    """
-    init_sentry()
-    # Run the server in stdio mode
-    logger.info("SecOps MCP server running on stdio")
-    mcp.run(show_banner=False)
-
-
-def run_http_with_uvicorn():
-    """Run the MCP server over HTTP using uvicorn ASGI server.
-
-    This is meant for local development. For production ready setup,
-    better use gunicorn with uvicorn ASGI workers via secops_mcp_server.http_app:app
-    """
-
-    init_sentry()
-    # Get host and port from environment variables
-    settings = get_settings()
-    mcp_port = int(settings.mcp_port or "8000")
-    mcp_host = settings.mcp_host
-
-    # Use StreamableHTTP transport with stateless JSON mode for scalability
-    import uvicorn
-
-    logger.info(f"Starting SecOps MCP server on {mcp_host}:{mcp_port}")
-    uvicorn.run(
-        mcp.http_app(path="/mcp", json_response=True, stateless_http=True),
-        host=mcp_host,
-        port=mcp_port,
-    )
+__all__ = ["run_http_with_uvicorn", "run_mcp_server", "run_stdio"]
 
 
 def run_mcp_server():
-    """Run the MCP server with transport auto-detection.
-
-    If MCP_PORT is set, uses StreamableHTTP transport.
-    Otherwise, uses stdio transport (default).
-    """
-    if get_settings().mcp_port:
-        run_http_with_uvicorn()
-    else:
-        run_stdio()
+    warnings.warn(
+        "secops-mcp-server is deprecated; use gg-mcp-server instead.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    _run_mcp_server()
 
 
 if __name__ == "__main__":
