+++ pyproject.toml
@@ -34,6 +34,7 @@ license = "MIT"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
+    "gg-mcp-server",
     "developer-mcp-server",
     "secops-mcp-server",
     "mcp~=1.24",  # CVE-2025-66416 / GHSA-9h52-p55h-vw2f: DNS rebinding vulnerability
@@ -55,6 +56,7 @@ dev = [
 ]
 
 [project.scripts]
+gg-mcp-server = "gg_mcp_server.run:run_mcp_server"
 developer-mcp-server = "developer_mcp_server.run:run_mcp_server"
 secops-mcp-server = "secops_mcp_server.run:run_mcp_server"
 
@@ -63,6 +65,7 @@ members = ["packages/*"]
 
 [tool.uv.sources]
 "gg-api-core" = { workspace = true }
+"gg-mcp-server" = { workspace = true }
 "developer-mcp-server" = { workspace = true }
 "secops-mcp-server" = { workspace = true }
