+++ scripts/run_http_server.py
@@ -8,11 +8,6 @@ os.environ["MCP_PORT"] = os.environ.get("MCP_PORT", "8088")
 os.environ["MCP_HOST"] = os.environ.get("MCP_HOST", "127.0.0.1")
 
 if __name__ == "__main__":
-    # For SecOps server:
-    from secops_mcp_server.run import run_mcp_server
+    from gg_mcp_server.run import run_mcp_server
 
-    # Or for Developer server:
-    # from developer_mcp_server.run import run_mcp_server
-
-    # Run the server
     run_mcp_server()
