+++ tests/test_server_profiles.py
@@ -1,8 +1,6 @@
-"""Test that server modules can be imported and initialized for both profiles.
+"""Test that the unified server module imports and registers tools correctly.
 
-This test ensures that both developer and secops server modules can be
-successfully imported and that their tool registration doesn't have syntax errors
-(like using mcp.add_tool instead of mcp.tool).
+Replaces the old per-profile tests now that there's a single gg_mcp_server.
 """
 
 import sys
@@ -13,10 +11,7 @@ import pytest
 
 @pytest.fixture
 def mock_env_no_http():
-    """Mock environment variables to prevent HTTP server from starting.
-
-    Sets ENABLE_LOCAL_OAUTH=true to use cached scope mode (stdio mode).
-    """
+    """Mock env vars to keep imports in stdio/cached-scope mode."""
     with patch.dict("os.environ", {"MCP_PORT": "", "ENABLE_LOCAL_OAUTH": "true"}, clear=False):
         yield
 
@@ -28,7 +23,6 @@ def mock_gitguardian_modules():
         mock_client = MagicMock()
         mock_client.get_current_token_info = AsyncMock(return_value={"scopes": ["scan"]})
         mock_get_client.return_value = mock_client
-
         yield {"get_client": mock_get_client}
 
 
@@ -39,135 +33,100 @@ def clean_module_imports(module_name: str):
         del sys.modules[module]
 
 
-class TestServerProfiles:
-    """Test that both server profiles can be initialized successfully."""
-
-    def test_developer_server_imports_successfully(self, mock_gitguardian_modules, mock_env_no_http):
-        """Test that the developer server module can be imported without errors.
-
-        This test would catch issues like:
-        - Using mcp.add_tool instead of mcp.tool
-        - Syntax errors in tool registration
-        - Missing imports
+class TestUnifiedServer:
+    def test_server_imports_successfully(self, mock_gitguardian_modules, mock_env_no_http):
         """
-        # Clean any previous imports
-        clean_module_imports("developer_mcp_server")
-
-        try:
-            # Import the developer server module
-            import developer_mcp_server.server as dev_server
-
-            # Verify the server was created
-            assert hasattr(dev_server, "mcp")
-            assert dev_server.mcp is not None
-
-            # Verify it's a GitGuardian MCP server (uses the abstract base class)
-            from gg_api_core.mcp_server import AbstractGitGuardianFastMCP
-
-            assert isinstance(dev_server.mcp, AbstractGitGuardianFastMCP)
-
-            # Verify the server has the expected name
-            assert dev_server.mcp.name == "GitGuardian Developer"
-
-        except AttributeError as e:
-            if "add_tool" in str(e):
-                pytest.fail(f"Developer server is using mcp.add_tool instead of mcp.tool: {e}")
-            raise
-        except Exception as e:
-            pytest.fail(f"Failed to import developer server: {e}")
-
-    def test_secops_server_imports_successfully(self, mock_gitguardian_modules, mock_env_no_http):
-        """Test that the secops server module can be imported without errors.
-
-        This test would catch issues like:
-        - Using mcp.add_tool instead of mcp.tool
-        - Syntax errors in tool registration
-        - Missing imports
+        GIVEN the unified gg_mcp_server package
+        WHEN its server module is imported
+        THEN an AbstractGitGuardianFastMCP instance is exposed as ``mcp``
         """
-        # Clean any previous imports
-        clean_module_imports("secops_mcp_server")
+        clean_module_imports("gg_mcp_server")
 
-        try:
-            # Import the secops server module
-            import secops_mcp_server.server as secops_server
+        import gg_mcp_server.server as srv
+        from gg_api_core.mcp_server import AbstractGitGuardianFastMCP
 
-            # Verify the server was created
-            assert hasattr(secops_server, "mcp")
-            assert secops_server.mcp is not None
-
-            # Verify it's a GitGuardian MCP server (uses the abstract base class)
-            from gg_api_core.mcp_server import AbstractGitGuardianFastMCP
-
-            assert isinstance(secops_server.mcp, AbstractGitGuardianFastMCP)
-
-            # Verify the server has the expected name
-            assert secops_server.mcp.name == "GitGuardian SecOps"
-
-        except AttributeError as e:
-            if "add_tool" in str(e):
-                pytest.fail(f"SecOps server is using mcp.add_tool instead of mcp.tool: {e}")
-            raise
-        except Exception as e:
-            pytest.fail(f"Failed to import secops server: {e}")
+        assert isinstance(srv.mcp, AbstractGitGuardianFastMCP)
+        assert srv.mcp.name == "GitGuardian"
 
     @pytest.mark.asyncio
-    async def test_developer_server_tools_registered(self, mock_gitguardian_modules, mock_env_no_http):
-        """Test that developer server has tools registered properly."""
-        # Clean any previous imports
-        clean_module_imports("developer_mcp_server")
+    async def test_secops_specific_tools_are_registered(self, mock_gitguardian_modules, mock_env_no_http):
+        """
+        GIVEN a token holding both read and write scopes
+        WHEN the unified server lists tools
+        THEN both developer-flavour and secops-flavour tools are present
+        """
+        clean_module_imports("gg_mcp_server")
 
-        import developer_mcp_server.server as dev_server
+        import gg_mcp_server.server as srv
 
-        # Mock the _fetch_token_scopes_from_api to avoid actual API calls
-        dev_server.mcp._fetch_token_scopes_from_api = AsyncMock()
-        dev_server.mcp._token_scopes = {"scan", "incidents:read"}
+        srv.mcp._fetch_token_scopes_from_api = AsyncMock()
+        srv.mcp._token_scopes = {
+            "scan",
+            "incidents:read",
+            "incidents:write",
+            "sources:read",
+            "honeytokens:read",
+            "honeytokens:write",
+        }
 
-        # List tools - this would fail if any tool was registered incorrectly
-        tools = await dev_server.mcp.list_tools()
+        tools = await srv.mcp.list_tools()
+        tool_names = {tool.name for tool in tools}
 
-        # Verify we have some tools registered
-        assert len(tools) > 0, "Developer server should have tools registered"
+        assert "list_incidents" in tool_names
+        assert "assign_incident" in tool_names
+        assert "create_code_fix_request" in tool_names
 
     @pytest.mark.asyncio
-    async def test_secops_server_tools_registered(self, mock_gitguardian_modules, mock_env_no_http):
-        """Test that secops server has tools registered properly."""
-        # Clean any previous imports
-        clean_module_imports("secops_mcp_server")
-
-        import secops_mcp_server.server as secops_server
-
-        # Mock the _fetch_token_scopes_from_api to avoid actual API calls
-        secops_server.mcp._fetch_token_scopes_from_api = AsyncMock()
-        secops_server.mcp._token_scopes = {"scan", "incidents:read", "incidents:write"}
-
-        # List tools - this would fail if any tool was registered incorrectly
-        tools = await secops_server.mcp.list_tools()
-
-        # Verify we have some tools registered
-        assert len(tools) > 0, "SecOps server should have tools registered"
-
-        # Verify some expected secops-specific tools are present
-        tool_names = [tool.name for tool in tools]
-        # Check for a secops-specific tool that should be registered
-        assert "assign_incident" in tool_names, "SecOps server should have assign_incident tool"
-        assert "create_code_fix_request" in tool_names, "SecOps server should have create_code_fix_request tool"
-
-    def test_both_servers_can_coexist(self, mock_gitguardian_modules, mock_env_no_http):
-        """Test that both server modules can be imported in the same test session.
-
-        This ensures there are no naming conflicts or import issues.
+    async def test_write_tools_hidden_without_write_scope(self, mock_gitguardian_modules, mock_env_no_http):
         """
-        # Clean any previous imports
+        GIVEN a token holding only read scopes
+        WHEN the unified server lists tools
+        THEN write tools are filtered out and read tools remain
+        """
+        clean_module_imports("gg_mcp_server")
+
+        import gg_mcp_server.server as srv
+
+        srv.mcp._fetch_token_scopes_from_api = AsyncMock()
+        srv.mcp._token_scopes = {"scan", "incidents:read", "sources:read"}
+
+        tools = await srv.mcp.list_tools()
+        tool_names = {tool.name for tool in tools}
+
+        assert "list_incidents" in tool_names
+        assert "assign_incident" not in tool_names
+        assert "create_code_fix_request" not in tool_names
+
+
+class TestDeprecatedShims:
+    def test_developer_shim_reexports_unified_server(self, mock_gitguardian_modules, mock_env_no_http):
+        """
+        GIVEN the deprecated developer_mcp_server.server shim
+        WHEN imported
+        THEN it re-exports the unified MCP instance and emits a DeprecationWarning
+        """
         clean_module_imports("developer_mcp_server")
+        clean_module_imports("gg_mcp_server")
+
+        with pytest.warns(DeprecationWarning):
+            import developer_mcp_server.server as shim
+
+        import gg_mcp_server.server as srv
+
+        assert shim.mcp is srv.mcp
+
+    def test_secops_shim_reexports_unified_server(self, mock_gitguardian_modules, mock_env_no_http):
+        """
+        GIVEN the deprecated secops_mcp_server.server shim
+        WHEN imported
+        THEN it re-exports the unified MCP instance and emits a DeprecationWarning
+        """
         clean_module_imports("secops_mcp_server")
+        clean_module_imports("gg_mcp_server")
 
-        try:
-            import developer_mcp_server.server as dev_server
-            import secops_mcp_server.server as secops_server
+        with pytest.warns(DeprecationWarning):
+            import secops_mcp_server.server as shim
 
-            # Both servers should be distinct instances
-            assert dev_server.mcp is not secops_server.mcp
-            assert dev_server.mcp.name != secops_server.mcp.name
+        import gg_mcp_server.server as srv
 
-        except Exception as e:
-            pytest.fail(f"Failed to import both servers: {e}")
+        assert shim.mcp is srv.mcp
